+++ src/components/UsersList/StyledComponents.ts
@@ -1,5 +1,4 @@
 import styled from 'styled-components/native';
-import CheckBox from '@react-native-community/checkbox';
 
 export const ScrollableContainer = styled.View`
   max-height: 100px;
@@ -15,13 +14,6 @@ export const UserItem = styled.TouchableHighlight`
   padding: 8px;
   border-bottom: 1px solid #f0f0f0;
   gap: 8px;
-`;
-
-
-export const Checkbox = styled(CheckBox)`
-  width: 16px;
-  height: 16px;
-  tint-color: #0052cd;
 `;
 
 export const Label = styled.Text`